requirements.txt
django>=1.6
setuptools-git
python-dateutil==2.7.5
mock
#python-openid
#django-openid-auth
pillow
